requirements.txt
openai
langchain
faiss-cpu
python-dotenv
tqdm
PyMuPDF
gradio
